Implement live server using vercel
