+++ client/src/components/stopwatch.js
@@ -125,7 +125,7 @@ function Stopwatch(props) {
                 }
             )
         })
-        axios.post("http://localhost:5000/home", {
+        axios.post("https://dowee2-server2.vercel.app/home", {
             headers:{
                 function: 'addNewHourlog'
             },newHourlog
@@ -141,7 +141,7 @@ function Stopwatch(props) {
 
     function setTimer(time){
         axios
-        .post("http://localhost:5000/home", {
+        .post("https://dowee2-server2.vercel.app/home", {
             headers:{
                 function: 'runTimer'
             },time
