+++ client/src/pages/home.js
@@ -23,7 +23,7 @@ const Home = () => {
 
     const memoizedFetchData = useCallback((userId) => {
         axios
-        .get('http://localhost:5000/home', {
+        .get('https://dowee2-server2.vercel.app/home', {
           headers:{ 
             userId: userId
           }
